Remove and ignore generated python and Django files
The source code contains *.pyc files which were produced by a python
while running an application. They are not needed in the repository
since they will be generated again anyway.

db.sqlite3 also not required since the "migrate" command will create a
database schema from the model.

emolab folder contains files of virtualenv. It is intended to create
virtualenv by a developer during deployment, and not stored in a
repository.